Created UserType.jsx and completed front end components
